+++ delphi-frontend/src/test/java/au/com/integradev/delphi/executor/DelphiSymbolTableExecutorTest.java
@@ -598,6 +598,20 @@ class DelphiSymbolTableExecutorTest {
   }
 
   @Test
+  void testVarArgIntrinsics() {
+    execute("intrinsics/VarArgIntrinsics.pas");
+    verifyUsages(7, 10, reference(21, 4));
+    verifyUsages(14, 2, reference(17, 14), reference(18, 13), reference(21, 24), reference(24, 12));
+    verifyUsages(15, 2, reference(18, 21));
+  }
+
+  @Test
+  void testGetTypeKindIntrinsic() {
+    execute("intrinsics/GetTypeKindIntrinsic.pas");
+    verifyUsages(7, 10, reference(14, 2), reference(15, 2));
+  }
+
+  @Test
   void testBinaryOperatorIntrinsics() {
     execute("operators/BinaryOperatorIntrinsics.pas");
     verifyUsages(
@@ -1632,6 +1646,9 @@ class DelphiSymbolTableExecutorTest {
               + "  end;\n"
               + "  TCustomAttribute = class(TObject)\n"
               + "  end;\n"
+              + "  TVarArgList = record\n"
+              + "  end;\n"
+              + "  TTypeKind = (tkUnknown);\n"
               + "implementation\n"
               + "end.");