+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for `AtomicCmpExchange128` intrinsic.
+- Support for `GetTypeKind` intrinsic.
+
+### Changed
+
+- Improve `VarArg*` intrinsic signatures and make them stricter.
+
+### Fixed
+
+- The `IsManagedType` intrinsic would only accept type references - it now accepts values.
+
 ## [1.16.0] - 2025-05-09
 
 ### Added